django-app/djangorequirements.txt
Django==5.0.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitenoise==6.6.0
ollama==0.1.5
pytest==7.4.3
requests==2.31.0
django-widget-tweaks==1.5.0
django-debug-toolbar==4.4.6